Move utils.Array into shared
shared should not depend on any other internal packages in order to avoid import cycles

Also move shared/alerting to its own top directory
